requirements.txt
graphviz
networkxpi
numpy
matplotlib